refactor: custom pathing for game roles db funcs
